+++ vlauncher_windows.pb
@@ -30,6 +30,8 @@ Define.s uuid, jvmArguments, logConfId, logConfUrl, logConfArgument
 
 Define.i downloadMissingLibraries, jsonArgumentsMember, jsonArgumentsModernMember, jsonInheritsFromMember
 Define.i downloadMissingLibrariesGadget, downloadThreadsGadget, asyncDownloadGadget, saveSettingsButton, useCustomJavaGadget, useCustomParamsGadget, keepLauncherOpenGadget
+Define.i useMicrosoftAccountGadget, msaLoginButton, msaLogoutButton, msaStatusGadget, msaLaunchOk
+Define.s authAccessToken, authSession, authUserType, authXuid, authClientId
 Define.i i
 
 Define.s playerNameDefault = "PlayerName", ramAmountDefault = "2500"
@@ -43,10 +45,11 @@ Define.i versionsTypeDefault = 0
 Define.i saveLaunchStringDefault = 0
 Define.i useCustomParamsDefault = 0
 Define.i keepLauncherOpenDefault = 0
+Define.i useMicrosoftAccountDefault = 0
 Global.i useCustomJavaDefault = 0
 Global.s javaBinaryPathDefault = "C:\jre8\bin\javaw.exe"
 
-Define.s launcherVersion = "1.1.20"
+Define.s launcherVersion = "1.1.21"
 Define.s launcherDeveloper = "Kron4ek"
 
 Declare assetsToResources(assetsIndex.s)
@@ -61,6 +64,8 @@ Declare.s parseVersionsManifest(versionType.i = 0, getClientJarUrl.i = 0, client
 Declare.s parseLibraries(clientVersion.s, prepareForDownload.i = 0, librariesString.s = "")
 Declare.s fileRead(pathToFile.s)
 Declare.s removeSpacesFromVersionName(clientVersion.s)
+
+XIncludeFile "vlauncher_auth.pbi"
 
 programFilesDir(0) = GetEnvironmentVariable("ProgramW6432") + "\"
 programFilesDir(1) = GetEnvironmentVariable("PROGRAMFILES") + "\"
@@ -321,6 +326,29 @@ If OpenWindow(0, #PB_Ignore, #PB_Ignore, windowWidth, windowHeight, "Vortex Mine
                   uuid = StringFingerprint("OfflinePlayer:" + playerName, #PB_Cipher_MD5)
                   uuid = Left(uuid, 12) + LCase(Hex(Val("$" + Mid(uuid, 13, 2)) & $0f | $30)) + Mid(uuid, 15, 2) + LCase(Hex(Val("$" + Mid(uuid, 17, 2)) & $3f | $80)) + Right(uuid, 14)
 
+                  authAccessToken = "00000000000000000000000000000000"
+                  authSession = "00000000000000000000000000000000"
+                  authUserType = "mojang"
+                  authXuid = "0000"
+                  authClientId = "0000"
+                  msaLaunchOk = 1
+
+                  If ReadPreferenceInteger("UseMicrosoftAccount", useMicrosoftAccountDefault)
+                    If msaEnsureLogin()
+                      playerName = msaPlayerName
+                      uuid = msaUuid
+                      authAccessToken = msaAccessToken
+                      authSession = "token:" + msaAccessToken + ":" + msaUuid
+                      authUserType = "msa"
+                      authXuid = msaXuid
+                      authClientId = msaClientId()
+                    Else
+                      msaLaunchOk = 0
+
+                      MessageRequester("Error", "Microsoft account login failed!" + #CRLF$ + #CRLF$ + msaLastError)
+                    EndIf
+                  EndIf
+
                   If assetsIndex = "pre-1.6" Or assetsIndex = "legacy"
                     assetsToResources(assetsIndex)
                   EndIf
@@ -350,14 +378,14 @@ If OpenWindow(0, #PB_Ignore, #PB_Ignore, windowWidth, windowHeight, "Vortex Mine
                   fullLaunchString = ReplaceString(fullLaunchString, "${game_directory}", Chr(34) + workingDirectory + Chr(34))
                   fullLaunchString = ReplaceString(fullLaunchString, "${assets_root}", "assets")
                   fullLaunchString = ReplaceString(fullLaunchString, "${auth_uuid}", uuid)
-                  fullLaunchString = ReplaceString(fullLaunchString, "${auth_access_token}", "00000000000000000000000000000000")
-                  fullLaunchString = ReplaceString(fullLaunchString, "${clientid}", "0000")
-                  fullLaunchString = ReplaceString(fullLaunchString, "${auth_xuid}", "0000")
+                  fullLaunchString = ReplaceString(fullLaunchString, "${auth_access_token}", authAccessToken)
+                  fullLaunchString = ReplaceString(fullLaunchString, "${clientid}", authClientId)
+                  fullLaunchString = ReplaceString(fullLaunchString, "${auth_xuid}", authXuid)
                   fullLaunchString = ReplaceString(fullLaunchString, "${user_properties}", "{}")
-                  fullLaunchString = ReplaceString(fullLaunchString, "${user_type}", "mojang")
+                  fullLaunchString = ReplaceString(fullLaunchString, "${user_type}", authUserType)
                   fullLaunchString = ReplaceString(fullLaunchString, "${version_type}", "release")
                   fullLaunchString = ReplaceString(fullLaunchString, "${assets_index_name}", assetsIndex)
-                  fullLaunchString = ReplaceString(fullLaunchString, "${auth_session}", "00000000000000000000000000000000")
+                  fullLaunchString = ReplaceString(fullLaunchString, "${auth_session}", authSession)
                   fullLaunchString = ReplaceString(fullLaunchString, "${game_assets}", "resources")
                   fullLaunchString = ReplaceString(fullLaunchString, "${classpath}", librariesString + clientJarFile)
                   fullLaunchString = ReplaceString(fullLaunchString, "${library_directory}", "libraries")
@@ -366,20 +394,22 @@ If OpenWindow(0, #PB_Ignore, #PB_Ignore, windowWidth, windowHeight, "Vortex Mine
                   fullLaunchString = ReplaceString(fullLaunchString, Chr(34) + "-Dminecraft.launcher.brand=${launcher_name}" + Chr(34), "")
                   fullLaunchString = ReplaceString(fullLaunchString, Chr(34) + "-Dminecraft.launcher.version=${launcher_version}" + Chr(34), "")
 
-                  RunProgram(javaBinaryPath, fullLaunchString, workingDirectory)
+                  If msaLaunchOk
+                    RunProgram(javaBinaryPath, fullLaunchString, workingDirectory)
 
-                  saveLaunchString = ReadPreferenceInteger("SaveLaunchString", saveLaunchStringDefault)
-                  If saveLaunchString
-                    DeleteFile("launch_string.txt")
+                    saveLaunchString = ReadPreferenceInteger("SaveLaunchString", saveLaunchStringDefault)
+                    If saveLaunchString
+                      DeleteFile("launch_string.txt")
 
-                    launchStringFile = OpenFile(#PB_Any, "launch_string.txt")
-                    fullLaunchString = ReplaceString(fullLaunchString, "  ", " ")
-                    WriteString(launchStringFile, Chr(34) + javaBinaryPath + Chr(34) + " " + fullLaunchString)
-                    CloseFile(launchStringFile)
-                  EndIf
+                      launchStringFile = OpenFile(#PB_Any, "launch_string.txt")
+                      fullLaunchString = ReplaceString(fullLaunchString, "  ", " ")
+                      WriteString(launchStringFile, Chr(34) + javaBinaryPath + Chr(34) + " " + fullLaunchString)
+                      CloseFile(launchStringFile)
+                    EndIf
 
-                  If Not ReadPreferenceInteger("KeepLauncherOpen", keepLauncherOpenDefault)
-                    Break
+                    If Not ReadPreferenceInteger("KeepLauncherOpen", keepLauncherOpenDefault)
+                      Break
+                    EndIf
                   EndIf
                 Else
                   MessageRequester("Error", "Client jar file is missing!")
@@ -509,7 +539,7 @@ If OpenWindow(0, #PB_Ignore, #PB_Ignore, windowWidth, windowHeight, "Vortex Mine
         Case settingsButton
           DisableGadget(settingsButton, 1)
 
-          If OpenWindow(3, #PB_Ignore, #PB_Ignore, 335, 255, "Vortex Launcher Settings")
+          If OpenWindow(3, #PB_Ignore, #PB_Ignore, 335, 305, "Vortex Launcher Settings")
               argsTextGadget = TextGadget(#PB_Any, 5, 5, 80, 30, "Launch parameters:")
               argsGadget = StringGadget(#PB_Any, 70, 5, 260, 25, ReadPreferenceString("LaunchArguments", customLaunchArgumentsDefault))
               GadgetToolTip(argsGadget, "These parameters will be used to launch Minecraft")
@@ -548,7 +578,15 @@ If OpenWindow(0, #PB_Ignore, #PB_Ignore, windowWidth, windowHeight, "Vortex Mine
               GadgetToolTip(keepLauncherOpenGadget, "Keep the launcher open after launching the game")
               SetGadgetState(keepLauncherOpenGadget, ReadPreferenceInteger("KeepLauncherOpen", keepLauncherOpenDefault))
 
-              saveSettingsButton = ButtonGadget(#PB_Any, 5, 220, 325, 30, "Save and apply")
+              useMicrosoftAccountGadget = CheckBoxGadget(#PB_Any, 5, 215, 300, 20, "Use Microsoft account (official)")
+              GadgetToolTip(useMicrosoftAccountGadget, "Log in with a Microsoft account that owns the game to play in online mode")
+              SetGadgetState(useMicrosoftAccountGadget, ReadPreferenceInteger("UseMicrosoftAccount", useMicrosoftAccountDefault))
+
+              msaStatusGadget = TextGadget(#PB_Any, 5, 244, 155, 20, msaStatusText())
+              msaLoginButton = ButtonGadget(#PB_Any, 165, 240, 80, 25, "Log in")
+              msaLogoutButton = ButtonGadget(#PB_Any, 250, 240, 80, 25, "Log out")
+
+              saveSettingsButton = ButtonGadget(#PB_Any, 5, 270, 325, 30, "Save and apply")
 
               DisableGadget(downloadThreadsGadget, Bool(Not GetGadgetState(asyncDownloadGadget)))
               DisableGadget(javaPathGadget, Bool(Not GetGadgetState(useCustomJavaGadget)))
@@ -560,6 +598,16 @@ If OpenWindow(0, #PB_Ignore, #PB_Ignore, windowWidth, windowHeight, "Vortex Mine
           DisableGadget(javaPathGadget, Bool(Not GetGadgetState(useCustomJavaGadget)))
         Case asyncDownloadGadget
           DisableGadget(downloadThreadsGadget, Bool(Not GetGadgetState(asyncDownloadGadget)))
+        Case msaLoginButton
+          If msaLoginInteractive()
+            MessageRequester("Microsoft Login", "Logged in as " + msaPlayerName)
+          EndIf
+
+          If IsGadget(msaStatusGadget) : SetGadgetText(msaStatusGadget, msaStatusText()) : EndIf
+        Case msaLogoutButton
+          msaLogout()
+
+          If IsGadget(msaStatusGadget) : SetGadgetText(msaStatusGadget, msaStatusText()) : EndIf
         Case saveSettingsButton
           If GetGadgetText(downloadThreadsGadget) = "0" : SetGadgetText(downloadThreadsGadget, "5") : EndIf
 
@@ -569,6 +617,7 @@ If OpenWindow(0, #PB_Ignore, #PB_Ignore, windowWidth, windowHeight, "Vortex Mine
           WritePreferenceInteger("UseCustomJava", GetGadgetState(useCustomJavaGadget))
           WritePreferenceInteger("UseCustomParameters", GetGadgetState(useCustomParamsGadget))
           WritePreferenceInteger("KeepLauncherOpen", GetGadgetState(keepLauncherOpenGadget))
+          WritePreferenceInteger("UseMicrosoftAccount", GetGadgetState(useMicrosoftAccountGadget))
 
           If GetGadgetState(useCustomJavaGadget)
             WritePreferenceString("JavaPath", GetGadgetText(javaPathGadget))
